HiGHS — Module

This module exists solely to collate the docstrings of libhighs.jl

Main.HiGHS.Highs_addCol — Method

Highs_addCol(highs, cost, lower, upper, num_new_nz, index, value)

Add a new column (variable) to the model.

Parameters

highs: a pointer to the Highs instance

cost: objective coefficient of the column

lower: lower bound of the column

upper: upper bound of the column

num_new_nz: number of non-zeros in the column

index: array of size [num_new_nz] with the row indices

value: array of size [num_new_nz] with row values

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_addCols — Method

Highs_addCols(highs, num_new_col, costs, lower, upper, num_new_nz, starts, index, value)

Add multiple columns (variables) to the model.

Parameters

highs: a pointer to the Highs instance

num_new_col: number of new columns to add

costs: array of size [num_new_col] with objective coefficients

lower: array of size [num_new_col] with lower bounds

upper: array of size [num_new_col] with upper bounds

num_new_nz: number of new nonzeros in the constraint matrix

starts: the constraint coefficients are given as a matrix in compressed sparse column form by the arrays starts, index, and value. starts is an array of size [num_new_cols] with the start index of each row in indices and values.

index: array of size [num_new_nz] with row indices

value: array of size [num_new_nz] with row values

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_addRow — Method

Highs_addRow(highs, lower, upper, num_new_nz, index, value)

Add a new row (a linear constraint) to the model.

Parameters

highs: a pointer to the Highs instance

lower: lower bound of the row

upper: upper bound of the row

num_new_nz: number of non-zeros in the row

index: array of size [num_new_nz] with column indices

value: array of size [num_new_nz] with column values

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_addRows — Method

Highs_addRows(highs, num_new_row, lower, upper, num_new_nz, starts, index, value)

Add multiple rows (linear constraints) to the model.

Parameters

highs: a pointer to the Highs instance

num_new_row: the number of new rows to add

lower: array of size [num_new_row] with the lower bounds of the rows

upper: array of size [num_new_row] with the upper bounds of the rows

num_new_nz: number of non-zeros in the rows

starts: the constraint coefficients are given as a matrix in compressed sparse row form by the arrays starts, index, and value. starts is an array of size [num_new_rows] with the start index of each row in indices and values.

index: array of size [num_new_nz] with column indices

value: array of size [num_new_nz] with column values

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_addVar — Method

Highs_addVar(highs, lower, upper)

Add a new variable to the model.

Parameters

highs: a pointer to the Highs instance

lower: lower bound of the column

upper: upper bound of the column

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_addVars — Method

Highs_addVars(highs, num_new_var, lower, upper)

Add multiple variables to the model.

Parameters

highs: a pointer to the Highs instance

num_new_var: number of new variables to add

lower: array of size [num_new_var] with lower bounds

upper: array of size [num_new_var] with upper bounds

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeCoeff — Method

Highs_changeCoeff(highs, row, col, value)

Change a coefficient in the constraint matrix.

Parameters

highs: a pointer to the Highs instance

row: the index of the row to change

col: the index of the col to change

value: the new constraint coefficient

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColBounds — Method

Highs_changeColBounds(highs, col, lower, upper)

Change the variable bounds of a column.

Parameters

highs: a pointer to the Highs instance

col: the index of the column whose bounds are to change

lower: the new lower bound

upper: the new upper bound

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColCost — Method

Highs_changeColCost(highs, col, cost)

Change the objective coefficient of a column.

Parameters

highs: a pointer to the Highs instance

col: the index of the column fo change

cost: the new objective coefficient

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColIntegrality — Method

Highs_changeColIntegrality(highs, col, integrality)

Change the integrality of a column.

Parameters

highs: a pointer to the Highs instance

col: the column index to change

integrality: the new integrality of the column in the form of a kHighsVarType constant

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColsBoundsByMask — Method

Highs_changeColsBoundsByMask(highs, mask, lower, upper)

Change the variable bounds of multiple columns given by a mask.

Parameters

highs: a pointer to the Highs instance

mask: an array of length [num_col] with 1 if the column bounds should be changed and 0 otherwise

lower: an array of length [num_col] with the new lower bounds

upper: an array of length [num_col] with the new upper bounds

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColsBoundsByRange — Method

Highs_changeColsBoundsByRange(highs, from_col, to_col, lower, upper)

Change the variable bounds of multiple adjacent columns.

Parameters

highs: a pointer to the Highs instance

from_col: the index of the first column whose bound changes

to_col: the index of the last column whose bound changes

lower: an array of length [to_col - from_col + 1] with the new lower bounds

upper: an array of length [to_col - from_col + 1] with the new upper bounds

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColsBoundsBySet — Method

Highs_changeColsBoundsBySet(highs, num_set_entries, set, lower, upper)

Change the bounds of multiple columns given by an array of indices.

Parameters

highs: a pointer to the Highs instance

num_set_entries: the number of columns to change

set: an array of size [num_set_entries] with the indices of the columns to change

lower: an array of length [num_set_entries] with the new lower bounds

upper: an array of length [num_set_entries] with the new upper bounds

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColsCostByMask — Method

Highs_changeColsCostByMask(highs, mask, cost)

Change the cost of multiple columns given by a mask.

Parameters

highs: a pointer to the Highs instance

mask: an array of length [num_col] with 1 if the column cost should be changed and 0 otherwise

cost: an array of length [num_col] with the new costs

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColsCostByRange — Method

Highs_changeColsCostByRange(highs, from_col, to_col, cost)

Change the cost coefficients of multiple adjacent columns.

Parameters

highs: a pointer to the Highs instance

from_col: the index of the first column whose cost changes

to_col: the index of the last column whose cost changes

cost: an array of length [to_col - from_col + 1] with the new objective coefficients

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColsCostBySet — Method

Highs_changeColsCostBySet(highs, num_set_entries, set, cost)

Change the cost of multiple columns given by an array of indices.

Parameters

highs: a pointer to the Highs instance

num_set_entries: the number of columns to change

set: an array of size [num_set_entries] with the indices of the columns to change

cost: an array of length [num_set_entries] with the new costs of the columns.

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColsIntegralityByMask — Method

Highs_changeColsIntegralityByMask(highs, mask, integrality)

Change the integrality of multiple columns given by a mask.

Parameters

highs: a pointer to the Highs instance

mask: an array of length [num_col] with 1 if the column integrality should be changed and 0 otherwise

integrality: an array of length [num_col] with the new integralities of the columns in the form of kHighsVarType constants

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColsIntegralityByRange — Method

Highs_changeColsIntegralityByRange(highs, from_col, to_col, integrality)

Change the integrality of multiple adjacent columns.

Parameters

highs: a pointer to the Highs instance

from_col: the index of the first column whose integrality changes

to_col: the index of the last column whose integrality changes

integrality: an array of length [to_col - from_col + 1] with the new integralities of the columns in the form of kHighsVarType constants

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeColsIntegralityBySet — Method

Highs_changeColsIntegralityBySet(highs, num_set_entries, set, integrality)

Change the integrality of multiple columns given by an array of indices.

Parameters

highs: a pointer to the Highs instance

num_set_entries: the number of columns to change

set: an array of size [num_set_entries] with the indices of the columns to change

integrality: an array of length [num_set_entries] with the new integralities of the columns in the form of kHighsVarType constants

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeObjectiveOffset — Method

Highs_changeObjectiveOffset(highs, offset)

Change the objective offset of the model.

Parameters

highs: a pointer to the Highs instance

offset: the new objective offset

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeObjectiveSense — Method

Highs_changeObjectiveSense(highs, sense)

Change the objective sense of the model.

Parameters

highs: a pointer to the Highs instance

sense: the new optimization sense in the form of a kHighsObjSense constant

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeRowBounds — Method

Highs_changeRowBounds(highs, row, lower, upper)

Change the bounds of a row.

Parameters

highs: a pointer to the Highs instance

row: the index of the row whose bounds are to change

lower: the new lower bound

upper: the new upper bound

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeRowsBoundsByMask — Method

Highs_changeRowsBoundsByMask(highs, mask, lower, upper)

Change the bounds of multiple rows given by a mask.

Parameters

highs: a pointer to the Highs instance

mask: an array of length [num_row] with 1 if the row bounds should be changed and 0 otherwise

lower: an array of length [num_row] with the new lower bounds

upper: an array of length [num_row] with the new upper bounds

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_changeRowsBoundsBySet — Method

Highs_changeRowsBoundsBySet(highs, num_set_entries, set, lower, upper)

Change the bounds of multiple rows given by an array of indices.

Parameters

highs: a pointer to the Highs instance

num_set_entries: the number of rows to change

set: an array of size [num_set_entries] with the indices of the rows to change

lower: an array of length [num_set_entries] with the new lower bounds

upper: an array of length [num_set_entries] with the new upper bounds

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_clear — Method

Highs_clear(highs)

Reset the options and then calls clearModel()

See Highs_destroy to free all associated memory.

Parameters

highs: a pointer to the Highs instance

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_clearModel — Method

Highs_clearModel(highs)

Remove all variables and constraints from the model highs, but do not invalidate the pointer highs. Future calls (for example, adding new variables and constraints) are allowed.

Parameters

highs: a pointer to the Highs instance

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_clearSolver — Method

Highs_clearSolver(highs)

Clear all solution data associated with the model

See Highs_destroy to clear the model and free all associated memory.

Parameters

highs: a pointer to the Highs instance

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_compilationDate — Method

Highs_compilationDate()

Return the HiGHS compilation date

Returns

the HiGHS compilation date

Main.HiGHS.Highs_create — Method

Highs_create()

Create a Highs instance and return the reference.

Call Highs_destroy on the returned reference to clean up allocated memory.

Returns

A pointer to the Highs instance

Main.HiGHS.Highs_crossover — Method

Highs_crossover(highs, num_col, num_row, col_value, col_dual, row_dual)

Set a primal (and possibly dual) solution as a starting point, then run crossover to compute a basic feasible solution. If there is no dual solution, pass col_dual and row_dual as nullptr.

Parameters

highs: a pointer to the Highs instance

num_col: the number of variables

num_row: the number of rows

col_value: array of length [num_col] with optimal primal solution for each column

col_dual: array of length [num_col] with optimal dual solution for each column

row_dual: array of length [num_row] with optimal dual solution for each row

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_deleteColsByMask — Method

Highs_deleteColsByMask(highs, mask)

Delete multiple columns given by a mask.

Parameters

highs: a pointer to the Highs instance

mask: an array of length [num_col] with 1 if the column should be deleted and 0 otherwise

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_deleteColsByRange — Method

Highs_deleteColsByRange(highs, from_col, to_col)

Delete multiple adjacent columns.

Parameters

highs: a pointer to the Highs instance

from_col: the index of the first column to delete

to_col: the index of the last column to delete

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_deleteColsBySet — Method

Highs_deleteColsBySet(highs, num_set_entries, set)

Delete multiple columns given by an array of indices.

Parameters

highs: a pointer to the Highs instance

num_set_entries: the number of columns to delete

set: an array of size [num_set_entries] with the indices of the columns to delete

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_deleteRowsByMask — Method

Highs_deleteRowsByMask(highs, mask)

Delete multiple rows given by a mask.

Parameters

highs: a pointer to the Highs instance

mask: an array of length [num_row] with 1 if the row should be deleted and 0 otherwise. New index of any column not deleted is returned in place of the value 0.

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_deleteRowsByRange — Method

Highs_deleteRowsByRange(highs, from_row, to_row)

Delete multiple adjacent rows.

Parameters

highs: a pointer to the Highs instance

from_row: the index of the first row to delete

to_row: the index of the last row to delete

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_deleteRowsBySet — Method

Highs_deleteRowsBySet(highs, num_set_entries, set)

Delete multiple rows given by an array of indices.

Parameters

highs: a pointer to the Highs instance

num_set_entries: the number of rows to delete

set: an array of size [num_set_entries] with the indices of the rows to delete

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_destroy — Method

Highs_destroy(highs)

Destroy the model highs created by Highs_create and free all corresponding memory. Future calls using highs are not allowed.

To empty a model without invalidating highs, see Highs_clearModel.

Parameters

highs: a pointer to the Highs instance

Main.HiGHS.Highs_getBasicVariables — Method

Highs_getBasicVariables(highs, basic_variables)

Get the indices of the rows and columns that make up the basis matrix of a basic feasible solution.

Non-negative entries are indices of columns, and negative entries are -row\_index - 1. For example, {1, -1} would be the second column and first row.

The order of these rows and columns is important for calls to the functions: - Highs_getBasisInverseRow - Highs_getBasisInverseCol - Highs_getBasisSolve - Highs_getBasisTransposeSolve - Highs_getReducedRow - Highs_getReducedColumn

Parameters

highs: a pointer to the Highs instance

basic_variables: array of size [num_rows], filled with the indices of the basic variables

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getBasis — Method

Highs_getBasis(highs, col_status, row_status)

Given a linear program with a basic feasible solution, get the column and row basis statuses.

Parameters

highs: a pointer to the Highs instance

col_status: array of length [num_col], to be filled with the column basis statuses in the form of a kHighsBasisStatus constant

row_status: array of length [num_row], to be filled with the row basis statuses in the form of a kHighsBasisStatus constant

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getBasisInverseCol — Method

Highs_getBasisInverseCol(highs, col, col_vector, col_num_nz, col_index)

Get a column of the inverse basis matrix

B^{-1}

.

See Highs_getBasicVariables for a description of the B matrix.

The arrays col_vector and col_index must have an allocated length of [num_row]. However, check col_num_nz to see how many non-zero elements are actually stored.

Parameters

highs: a pointer to the Highs instance

col: index of the column to compute

col_vector: values of the non-zero elements

col_num_nz: the number of non-zeros in the column

col_index: indices of the non-zero elements

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getBasisInverseRow — Method

Highs_getBasisInverseRow(highs, row, row_vector, row_num_nz, row_index)

Get a row of the inverse basis matrix

B^{-1}

.

See Highs_getBasicVariables for a description of the B matrix.

The arrays row_vector and row_index must have an allocated length of [num_row]. However, check row_num_nz to see how many non-zero elements are actually stored.

Parameters

highs: a pointer to the Highs instance

row: index of the row to compute

row_vector: values of the non-zero elements

row_num_nz: the number of non-zeros in the row

row_index: indices of the non-zero elements

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getBasisSolve — Method

Highs_getBasisSolve(highs, rhs, solution_vector, solution_num_nz, solution_index)

Compute

\mathbf{x}=B^{-1}\mathbf{b}

for a given vector

\mathbf{b}

.

See Highs_getBasicVariables for a description of the B matrix.

The arrays solution_vector and solution_index must have an allocated length of [num_row]. However, check solution_num_nz to see how many non-zero elements are actually stored.

Parameters

highs: a pointer to the Highs instance

rhs: the right-hand side vector b

solution_vector: values of the non-zero elements

solution_num_nz: the number of non-zeros in the solution

solution_index: indices of the non-zero elements

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getBasisTransposeSolve — Method

Highs_getBasisTransposeSolve(highs, rhs, solution_vector, solution_nz, solution_index)

Compute

\mathbf{x}=B^{-T}\mathbf{b}

for a given vector

\mathbf{b}

.

See Highs_getBasicVariables for a description of the B matrix.

The arrays solution_vector and solution_index must have an allocated length of [num_row]. However, check solution_num_nz to see how many non-zero elements are actually stored.

Parameters

highs: a pointer to the Highs instance

rhs: the right-hand side vector b

solution_vector: values of the non-zero elements

solution_num_nz: the number of non-zeros in the solution

solution_index: indices of the non-zero elements

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getBoolOptionValue — Method

Highs_getBoolOptionValue(highs, option, value)

Get a boolean-valued option.

Parameters

highs: a pointer to the Highs instance

option: the name of the option

value: storage for the value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getBoolOptionValues — Method

Highs_getBoolOptionValues(highs, option, current_value, default_value)

Get the current and default values of a bool option

Parameters

highs: a pointer to the Highs instance

current_value: a pointer to the current value of the option

default_value: a pointer to the default value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getColIntegrality — Method

Highs_getColIntegrality(highs, col, integrality)

Get the integrality of a column

Parameters

col: the column for which the name is required

integrality: the integrality of the column

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getColName — Method

Highs_getColName(highs, col, name)

Get the name of a column

Parameters

col: the column for which the name is required

name: the name of the column

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getColsByMask — Method

Highs_getColsByMask(highs, mask, num_col, costs, lower, upper, num_nz, matrix_start, matrix_index, matrix_value)

Get data associated with multiple columns given by a mask.

This function is identical to Highs_getColsByRange, except for how the columns are specified.

Parameters

mask: array of length [num_col] containing a 1 to get the column and 0 otherwise

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getColsByRange — Method

Highs_getColsByRange(highs, from_col, to_col, num_col, costs, lower, upper, num_nz, matrix_start, matrix_index, matrix_value)

Get data associated with multiple adjacent columns from the model.

To query the constraint coefficients, this function should be called twice: - First, call this function with matrix_start, matrix_index, and matrix_value as NULL. This call will populate num_nz with the number of nonzero elements in the corresponding section of the constraint matrix. - Second, allocate new matrix_index and matrix_value arrays of length num_nz and call this function again to populate the new arrays with their contents.

Parameters

highs: a pointer to the Highs instance

from_col: the first column for which to query data for

to_col: the last column (inclusive) for which to query data for

num_col: an integer populated with the number of columns got from the model (this should equal to\_col - from\_col + 1)

costs: array of size [to_col - from_col + 1] for the column cost coefficients

lower: array of size [to_col - from_col + 1] for the column lower bounds

upper: array of size [to_col - from_col + 1] for the column upper bounds

num_nz: an integer populated with the number of non-zero elements in the constraint matrix

matrix_start: array of size [to_col - from_col + 1] with the start indices of each column in matrix_index and matrix_value

matrix_index: array of size [num_nz] with the row indices of each element in the constraint matrix

matrix_value: array of size [num_nz] with the non-zero elements of the constraint matrix.

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getColsBySet — Method

Highs_getColsBySet(highs, num_set_entries, set, num_col, costs, lower, upper, num_nz, matrix_start, matrix_index, matrix_value)

Get data associated with multiple columns given by an array.

This function is identical to Highs_getColsByRange, except for how the columns are specified.

Parameters

num_set_indices: the number of indices in the set

set: array of size [num_set_entries] with the column indices to get

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getDoubleInfoValue — Method

Highs_getDoubleInfoValue(highs, info, value)

Get a double-valued info value.

Parameters

highs: a pointer to the Highs instance

info: the name of the info item

value: a reference to an double that the result will be stored in

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getDoubleOptionValue — Method

Highs_getDoubleOptionValue(highs, option, value)

Get a double-valued option.

Parameters

highs: a pointer to the Highs instance

option: the name of the option

value: storage for the value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getDoubleOptionValues — Method

Highs_getDoubleOptionValues(highs, option, current_value, min_value, max_value, default_value)

Get the current and default values of a double option

Parameters

highs: a pointer to the Highs instance

current_value: a pointer to the current value of the option

min_value: a pointer to the minimum value of the option

max_value: a pointer to the maximum value of the option

default_value: a pointer to the default value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getDualRay — Method

Highs_getDualRay(highs, has_dual_ray, dual_ray_value)

Get an unbounded dual ray that is a certificate of primal infeasibility.

Parameters

highs: a pointer to the Highs instance

has_dual_ray: a pointer to an int to store 1 if the dual ray exists

dual_ray_value: an array of length [num_row] filled with the unbounded ray

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getHessianNumNz — Method

Highs_getHessianNumNz(highs)

Return the number of nonzeroes in the Hessian matrix of the model.

Parameters

highs: a pointer to the Highs instance

Returns

the number of nonzeroes in the Hessian matrix of the model.

Main.HiGHS.Highs_getInfinity — Method

Highs_getInfinity(highs)

Return the value of infinity used by HiGHS.

Parameters

highs: a pointer to the Highs instance

Returns

the value of infinity used by HiGHS

Main.HiGHS.Highs_getInfoType — Method

Highs_getInfoType(highs, info, type)

Get the type expected by an info item.

Parameters

highs: a pointer to the Highs instance

info: the name of the info item

type: int in which the corresponding kHighsOptionType constant is stored

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getInt64InfoValue — Method

Highs_getInt64InfoValue(highs, info, value)

Get an int64-valued info value.

Parameters

highs: a pointer to the Highs instance

info: the name of the info item

value: a reference to a int64 that the result will be stored in

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getIntInfoValue — Method

Highs_getIntInfoValue(highs, info, value)

Get an int-valued info value.

Parameters

highs: a pointer to the Highs instance

info: the name of the info item

value: a reference to an integer that the result will be stored in

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getIntOptionValue — Method

Highs_getIntOptionValue(highs, option, value)

Get an int-valued option.

Parameters

highs: a pointer to the Highs instance

option: the name of the option

value: storage for the value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getIntOptionValues — Method

Highs_getIntOptionValues(highs, option, current_value, min_value, max_value, default_value)

Get the current and default values of an int option

Parameters

highs: a pointer to the Highs instance

current_value: a pointer to the current value of the option

min_value: a pointer to the minimum value of the option

max_value: a pointer to the maximum value of the option

default_value: a pointer to the default value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getModel — Method

Highs_getModel(highs, a_format, q_format, num_col, num_row, num_nz, hessian_num_nz, sense, offset, col_cost, col_lower, col_upper, row_lower, row_upper, a_start, a_index, a_value, q_start, q_index, q_value, integrality)

Get the data from a HiGHS model.

The input arguments have the same meaning (in a different order) to those used in Highs_passModel.

Note that all arrays must be pre-allocated to the correct size before calling Highs_getModel. Use the following query methods to check the appropriate size: - Highs_getNumCol - Highs_getNumRow - Highs_getNumNz - Highs_getHessianNumNz

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getModelStatus — Method

Highs_getModelStatus(highs)

Return the optimization status of the model in the form of a kHighsModelStatus constant.

Parameters

highs: a pointer to the Highs instance

Returns

an integer corresponding to the kHighsModelStatus constant

Main.HiGHS.Highs_getNumCol — Method

Highs_getNumCol(highs)

Return the number of columns in the model.

Parameters

highs: a pointer to the Highs instance

Returns

the number of columns in the model

Main.HiGHS.Highs_getNumNz — Method

Highs_getNumNz(highs)

Return the number of nonzeros in the constraint matrix of the model.

Parameters

highs: a pointer to the Highs instance

Returns

the number of nonzeros in the constraint matrix of the model.

Main.HiGHS.Highs_getNumOptions — Method

Highs_getNumOptions(highs)

Return the number of options

Parameters

highs: a pointer to the Highs instance

Main.HiGHS.Highs_getNumRow — Method

Highs_getNumRow(highs)

Return the number of rows in the model.

Parameters

highs: a pointer to the Highs instance

Returns

the number of rows in the model.

Main.HiGHS.Highs_getObjectiveOffset — Method

Highs_getObjectiveOffset(highs, offset)

Get the objective offset.

Parameters

highs: a pointer to the Highs instance

offset: stores the current objective offset

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getObjectiveSense — Method

Highs_getObjectiveSense(highs, sense)

Get the objective sense.

Parameters

highs: a pointer to the Highs instance

sense: stores the current objective sense as a kHighsObjSense constant

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getObjectiveValue — Method

Highs_getObjectiveValue(highs)

Get the primal objective function value.

Parameters

highs: a pointer to the Highs instance

Returns

the primal objective function value

Main.HiGHS.Highs_getOptionName — Method

Highs_getOptionName(highs, index, name)

Get the name of an option identified by index

Parameters

highs: a pointer to the Highs instance

index: the index of the option

name: the name of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getOptionType — Method

Highs_getOptionType(highs, option, type)

Get the type expected by an option.

Parameters

highs: a pointer to the Highs instance

option: the name of the option

type: int in which the corresponding kHighsOptionType constant is stored

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getPrimalRay — Method

Highs_getPrimalRay(highs, has_primal_ray, primal_ray_value)

Get an unbounded primal ray that is a certificate of dual infeasibility.

Parameters

highs: a pointer to the Highs instance

has_primal_ray: a pointer to an int to store 1 if the primal ray exists

primal_ray_value: an array of length [num_col] filled with the unbounded ray

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getReducedColumn — Method

Highs_getReducedColumn(highs, col, col_vector, col_num_nz, col_index)

Compute a column of

B^{-1}A

.

See Highs_getBasicVariables for a description of the B matrix.

The arrays col_vector and col_index must have an allocated length of [num_row]. However, check col_num_nz to see how many non-zero elements are actually stored.

Parameters

highs: a pointer to the Highs instance

col: index of the column to compute

col_vector: values of the non-zero elements

col_num_nz: the number of non-zeros in the column

col_index: indices of the non-zero elements

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getReducedRow — Method

Highs_getReducedRow(highs, row, row_vector, row_num_nz, row_index)

Compute a row of

B^{-1}A

.

See Highs_getBasicVariables for a description of the B matrix.

The arrays row_vector and row_index must have an allocated length of [num_row]. However, check row_num_nz to see how many non-zero elements are actually stored.

Parameters

highs: a pointer to the Highs instance

row: index of the row to compute

row_vector: values of the non-zero elements

row_num_nz: the number of non-zeros in the row

row_index: indices of the non-zero elements

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getRowName — Method

Highs_getRowName(highs, row, name)

Get the name of a row

Parameters

row: the row for which the name is required

name: the name of the row

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getRowsByMask — Method

Highs_getRowsByMask(highs, mask, num_row, lower, upper, num_nz, matrix_start, matrix_index, matrix_value)

Get data associated with multiple rows given by a mask.

This function is identical to Highs_getRowsByRange, except for how the rows are specified.

Parameters

mask: array of length [num_row] containing a 1 to get the row and 0 otherwise

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getRowsByRange — Method

Highs_getRowsByRange(highs, from_row, to_row, num_row, lower, upper, num_nz, matrix_start, matrix_index, matrix_value)

Get data associated with multiple adjacent rows from the model.

To query the constraint coefficients, this function should be called twice: - First, call this function with matrix_start, matrix_index, and matrix_value as NULL. This call will populate num_nz with the number of nonzero elements in the corresponding section of the constraint matrix. - Second, allocate new matrix_index and matrix_value arrays of length num_nz and call this function again to populate the new arrays with their contents.

Parameters

highs: a pointer to the Highs instance

from_row: the first row for which to query data for

to_row: the last row (inclusive) for which to query data for

num_row: an integer populated with the number of row got from the model

lower: array of size [to_row - from_row + 1] for the row lower bounds

upper: array of size [to_row - from_row + 1] for the row upper bounds

num_nz: an integer populated with the number of non-zero elements in the constraint matrix

matrix_start: array of size [to_row - from_row + 1] with the start indices of each row in matrix_index and matrix_value

matrix_index: array of size [num_nz] with the column indices of each element in the constraint matrix

matrix_value: array of size [num_nz] with the non-zero elements of the constraint matrix.

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getRowsBySet — Method

Highs_getRowsBySet(highs, num_set_entries, set, num_row, lower, upper, num_nz, matrix_start, matrix_index, matrix_value)

Get data associated with multiple rows given by an array.

This function is identical to Highs_getRowsByRange, except for how the rows are specified.

Parameters

num_set_indices: the number of indices in the set

set: array of size [num_set_entries] with the row indices to get

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getRunTime — Method

Highs_getRunTime(highs)

Return the cumulative wall-clock time spent in Highs_run.

Parameters

highs: a pointer to the Highs instance

Returns

the cumulative wall-clock time spent in Highs_run

Main.HiGHS.Highs_getSolution — Method

Highs_getSolution(highs, col_value, col_dual, row_value, row_dual)

Get the primal and dual solution from an optimized model.

Parameters

highs: a pointer to the Highs instance

col_value: array of length [num_col], filled with primal column values

col_dual: array of length [num_col], filled with dual column values

row_value: array of length [num_row], filled with primal row values

row_dual: array of length [num_row], filled with dual row values

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getStringOptionValue — Method

Highs_getStringOptionValue(highs, option, value)

Get a string-valued option.

Parameters

highs: a pointer to the Highs instance

option: the name of the option

value: pointer to allocated memory to store the value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_getStringOptionValues — Method

Highs_getStringOptionValues(highs, option, current_value, default_value)

Get the current and default values of a string option

Parameters

highs: a pointer to the Highs instance

current_value: a pointer to the current value of the option

default_value: a pointer to the default value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_githash — Method

Highs_githash()

Return the HiGHS githash

Returns

the HiGHS githash

Main.HiGHS.Highs_lpCall — Method

Highs_lpCall(num_col, num_row, num_nz, a_format, sense, offset, col_cost, col_lower, col_upper, row_lower, row_upper, a_start, a_index, a_value, col_value, col_dual, row_value, row_dual, col_basis_status, row_basis_status, model_status)

Formulate and solve a linear program using HiGHS.

Parameters

num_col: the number of columns

num_row: the number of rows

num_nz: the number of nonzeros in the constraint matrix

a_format: the format of the constraint matrix as a kHighsMatrixFormat constant

sense: the optimization sense as a kHighsObjSense constant

offset: the objective constant

col_cost: array of length [num_col] with the column costs

col_lower: array of length [num_col] with the column lower bounds

col_upper: array of length [num_col] with the column upper bounds

row_lower: array of length [num_row] with the row lower bounds

row_upper: array of length [num_row] with the row upper bounds

a_start: the constraint matrix is provided to HiGHS in compressed sparse column form (if a_format is kHighsMatrixFormatColwise, otherwise compressed sparse row form). The sparse matrix consists of three arrays, a_start, a_index, and a_value. a_start is an array of length [num_col] containing the starting index of each column in a_index. If a_format is kHighsMatrixFormatRowwise the array is of length [num_row] corresponding to each row.

a_index: array of length [num_nz] with indices of matrix entries

a_value: array of length [num_nz] with values of matrix entries

col_value: array of length [num_col], filled with the primal column solution

col_dual: array of length [num_col], filled with the dual column solution

row_value: array of length [num_row], filled with the primal row solution

row_dual: array of length [num_row], filled with the dual row solution

col_basis_status: array of length [num_col], filled with the basis status of the columns in the form of a kHighsBasisStatus constant

row_basis_status: array of length [num_row], filled with the basis status of the rows in the form of a kHighsBasisStatus constant

model_status: termination status of the model after the solve in the form of a kHighsModelStatus constant

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_mipCall — Method

Highs_mipCall(num_col, num_row, num_nz, a_format, sense, offset, col_cost, col_lower, col_upper, row_lower, row_upper, a_start, a_index, a_value, integrality, col_value, row_value, model_status)

Formulate and solve a mixed-integer linear program using HiGHS.

The signature of this method is identical to Highs_lpCall, except that it has an additional integrality argument, and that it is missing the col_dual, row_dual, col_basis_status and row_basis_status arguments.

Parameters

integrality: array of length [num_col] containing a kHighsVarType constant for each column

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_passColName — Method

Highs_passColName(highs, col, name)

Pass the name of a column

Parameters

col: the column for which the name is supplied

name: the name of the column

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_passHessian — Method

Highs_passHessian(highs, dim, num_nz, format, start, index, value)

Set the Hessian matrix for a quadratic objective.

Parameters

highs: a pointer to the Highs instance

dim: the dimension of the Hessian matrix. Should be [num_col].

num_nz: the number of non-zero elements in the Hessian matrix

format: the format of the Hessian matrix as a kHighsHessianFormat constant. This must be kHighsHessianFormatTriangular.

start: the Hessian matrix is provided to HiGHS as the upper triangular component in compressed sparse column form. The sparse matrix consists of three arrays, start, index, and value. start is an array of length [num_col] containing the starting index of each column in index.

index: array of length [num_nz] with indices of matrix entries

value: array of length [num_nz] with values of matrix entries

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_passLp — Method

Highs_passLp(highs, num_col, num_row, num_nz, a_format, sense, offset, col_cost, col_lower, col_upper, row_lower, row_upper, a_start, a_index, a_value)

Pass a linear program (LP) to HiGHS in a single function call.

The signature of this function is identical to Highs_passModel, without the arguments for passing the Hessian matrix of a quadratic program and the integrality vector.

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_passMip — Method

Highs_passMip(highs, num_col, num_row, num_nz, a_format, sense, offset, col_cost, col_lower, col_upper, row_lower, row_upper, a_start, a_index, a_value, integrality)

Pass a mixed-integer linear program (MILP) to HiGHS in a single function call.

The signature of function is identical to Highs_passModel, without the arguments for passing the Hessian matrix of a quadratic program.

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_passModel — Method

Highs_passModel(highs, num_col, num_row, num_nz, q_num_nz, a_format, q_format, sense, offset, col_cost, col_lower, col_upper, row_lower, row_upper, a_start, a_index, a_value, q_start, q_index, q_value, integrality)

Pass a model to HiGHS in a single function call. This is faster than constructing the model using Highs_addRow and Highs_addCol.

Parameters

highs: a pointer to the Highs instance

num_col: the number of columns

num_row: the number of rows

num_nz: the number of elements in the constraint matrix

q_num_nz: the number of elements in the Hessian matrix

a_format: the format of the constraint matrix to use in th form of a kHighsMatrixFormat constant

q_format: the format of the Hessian matrix to use in the form of a kHighsHessianFormat constant

sense: the optimization sense in the form of a kHighsObjSense constant

offset: the constant term in the objective function

col_cost: array of length [num_col] with the objective coefficients

col_lower: array of length [num_col] with the lower column bounds

col_upper: array of length [num_col] with the upper column bounds

row_lower: array of length [num_row] with the upper row bounds

row_upper: array of length [num_row] with the upper row bounds

a_start: the constraint matrix is provided to HiGHS in compressed sparse column form (if a_format is kHighsMatrixFormatColwise, otherwise compressed sparse row form). The sparse matrix consists of three arrays, a_start, a_index, and a_value. a_start is an array of length [num_col] containing the starting index of each column in a_index. If a_format is kHighsMatrixFormatRowwise the array is of length [num_row] corresponding to each row.

a_index: array of length [num_nz] with indices of matrix entries

a_value: array of length [num_nz] with values of matrix entries

q_start: the Hessian matrix is provided in the same format as the constraint matrix, using q_start, q_index, and q_value in the place of a_start, a_index, and a_value. If the model is linear, pass NULL.

q_index: array of length [q_num_nz] with indices of matrix entries. If the model is linear, pass NULL.

q_value: array of length [q_num_nz] with values of matrix entries. If the model is linear, pass NULL.

integrality: an array of length [num_col] containing a kHighsVarType consatnt for each column

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_passRowName — Method

Highs_passRowName(highs, row, name)

Pass the name of a row

Parameters

row: the row for which the name is supplied

name: the name of the row

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_qpCall — Method

Highs_qpCall(num_col, num_row, num_nz, q_num_nz, a_format, q_format, sense, offset, col_cost, col_lower, col_upper, row_lower, row_upper, a_start, a_index, a_value, q_start, q_index, q_value, col_value, col_dual, row_value, row_dual, col_basis_status, row_basis_status, model_status)

Formulate and solve a quadratic program using HiGHS.

The signature of this method is identical to Highs_lpCall, except that it has additional arguments for specifying the Hessian matrix.

Parameters

q_num_nz: the number of nonzeros in the Hessian matrix

q_format: the format of the Hessian matrix in the form of a kHighsHessianStatus constant. If q_num_nz > 0, this must be kHighsHessianFormatTriangular

q_start: the Hessian matrix is provided in the same format as the constraint matrix, using q_start, q_index, and q_value in the place of a_start, a_index, and a_value

q_index: array of length [q_num_nz] with indices of matrix entries

q_value: array of length [q_num_nz] with values of matrix entries

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_readModel — Method

Highs_readModel(highs, filename)

Read a model from filename into highs.

Parameters

highs: a pointer to the Highs instance

filename: the filename to read

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_resetGlobalScheduler — Method

Highs_resetGlobalScheduler(blocking)

Releases all resources held by the global scheduler instance. It is not thread-safe to call this function while calling Highs_run()/Highs_call() on any other Highs instance in any thread. After this function has terminated it is guaranteed that eventually all previously created scheduler threads will terminate and allocated memory will be released. After this function has returned the option value for the number of threads may be altered to a new value before the next call to Highs_run()/Highs_call(). If the given parameter has a nonzero value, then the function will not return until all memory is freed, which might be desirable when debugging heap memory but requires the calling thread to wait for all scheduler threads to wake-up which is usually not necessary.

Returns

No status is returned since the function call cannot fail. Calling this function while any Highs instance is in use on any thread is undefined behavior and may cause crashes, but cannot be detected and hence is fully in the callers responsibility.

Main.HiGHS.Highs_resetOptions — Method

Highs_resetOptions(highs)

Reset all options to their default value.

Parameters

highs: a pointer to the Highs instance

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_run — Method

Highs_run(highs)

Optimize a model. The algorithm used by HiGHS depends on the options that have been set.

Parameters

highs: a pointer to the Highs instance

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_scaleCol — Method

Highs_scaleCol(highs, col, scaleval)

Scale a column by a constant.

Scaling a column modifies the elements in the constraint matrix, the variable bounds, and the objective coefficient.

If scaleval < 0, the variable bounds flipped.

Parameters

highs: a pointer to the Highs instance

col: the index of the column to scale

scaleval: the value by which to scale the column

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_scaleRow — Method

Highs_scaleRow(highs, row, scaleval)

Scale a row by a constant.

If scaleval < 0, the row bounds are flipped.

Parameters

highs: a pointer to the Highs instance

row: the index of the row to scale

scaleval: the value by which to scale the row

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_setBasis — Method

Highs_setBasis(highs, col_status, row_status)

Set a basic feasible solution by passing the column and row basis statuses to the model.

Parameters

highs: a pointer to the Highs instance

col_status: an array of length [num_col] with the column basis status in the form of kHighsBasisStatus constants

row_status: an array of length [num_row] with the row basis status in the form of kHighsBasisStatus constants

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_setBoolOptionValue — Method

Highs_setBoolOptionValue(highs, option, value)

Set a boolean-valued option.

Parameters

highs: a pointer to the Highs instance

option: the name of the option

value: the value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_setDoubleOptionValue — Method

Highs_setDoubleOptionValue(highs, option, value)

Set a double-valued option.

Parameters

highs: a pointer to the Highs instance

option: the name of the option

value: the value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_setIntOptionValue — Method

Highs_setIntOptionValue(highs, option, value)

Set an int-valued option.

Parameters

highs: a pointer to the Highs instance

option: the name of the option

value: the value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_setLogicalBasis — Method

Highs_setLogicalBasis(highs)

Set a logical basis in the model.

Parameters

highs: a pointer to the Highs instance

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_setSolution — Method

Highs_setSolution(highs, col_value, row_value, col_dual, row_dual)

Set a solution by passing the column and row primal and dual solution values. For any values that are unavailable pass NULL.

Parameters

highs: a pointer to the Highs instance

col_value: an array of length [num_col] with the column solution values

row_value: an array of length [num_row] with the row solution values

col_dual: an array of length [num_col] with the column dual values

row_dual: an array of length [num_row] with the row dual values

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_setStringOptionValue — Method

Highs_setStringOptionValue(highs, option, value)

Set a string-valued option.

Parameters

highs: a pointer to the Highs instance

option: the name of the option

value: the value of the option

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_version — Method

Highs_version()

Return the HiGHS version number vX.Y.Z

Returns

the HiGHS version as a char*

Main.HiGHS.Highs_versionMajor — Method

Highs_versionMajor()

Return the HiGHS major version number

Returns

the HiGHS major version number

Main.HiGHS.Highs_versionMinor — Method

Highs_versionMinor()

Return the HiGHS minor version number

Returns

the HiGHS minor version number

Main.HiGHS.Highs_versionPatch — Method

Highs_versionPatch()

Return the HiGHS patch version number

Returns

the HiGHS patch version number

Main.HiGHS.Highs_writeModel — Method

Highs_writeModel(highs, filename)

Write the model in highs to filename.

Parameters

highs: a pointer to the Highs instance

filename: the filename to write.

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_writeOptions — Method

Highs_writeOptions(highs, filename)

Write the current options to file.

Parameters

highs: a pointer to the Highs instance

filename: the filename to write the options to

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_writeOptionsDeviations — Method

Highs_writeOptionsDeviations(highs, filename)

Write the value of non-default options to file.

This is similar to Highs_writeOptions, except only options with non-default value are written to filename.

Parameters

highs: a pointer to the Highs instance

filename: the filename to write the options to

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_writeSolution — Method

Highs_writeSolution(highs, filename)

Write the solution information (including dual and basis status, if available) to a file.

See also: Highs_writeSolutionPretty.

Parameters

highs: a pointer to the Highs instance

filename: the name of the file to write the results to

Returns

a kHighsStatus constant indicating whether the call succeeded

Main.HiGHS.Highs_writeSolutionPretty — Method

Highs_writeSolutionPretty(highs, filename)

Write the solution information (including dual and basis status, if available) to a file in a human-readable format.

The method identical to Highs_writeSolution, except that the printout is in a human-readiable format.

Parameters

highs: a pointer to the Highs instance

filename: the name of the file to write the results to

Returns

a kHighsStatus constant indicating whether the call succeeded